requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.3
Pillow==10.4.0